backend/requirements.txt
flask
flask-socketio
pymongo
pyjwt
flask-bcrypt
python-dotenv
flask-cors